scripts/gitlab_runner_pre_build/requirements.txt
boto3
pyjwt
requests